Initial commit of Go web crawler
My first Go project so I would expect a few things to be messy or
untidy.

Web crawlers are a bit messy because, well so is the internet, so I'd
expect a bunch of assumptions would need to change if this was used in
anger in the wild.

Added some tests for the parsing but not for the program in general
(yet?)